Update create.sql
+ change SMALLINT TO INT

+++ postgre/create.sql
@@ -8,24 +8,24 @@ DROP TABLE IF EXISTS Transaksi, PendaftaranToJadwal;
 DROP TYPE IF EXISTS GENDER, ACTIVE, GOL_DARAH,ROLE_P, STATUS_DAFTAR,HARI,METODE;
 
 CREATE TABLE  Kecamatan(
-    id_kecamatan SMALLINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id_kecamatan INT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     nama_kecamatan VARCHAR(20)
 );
 
 CREATE TABLE Kelurahan(
-    id_kelurahan SMALLINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id_kelurahan INT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     nama_kelurahan VARCHAR(20),
-    id_kecamatan SMALLINT REFERENCES Kecamatan(id_kecamatan)
+    id_kecamatan INT REFERENCES Kecamatan(id_kecamatan)
 );
 
 CREATE TABLE Spesialisasi(
-    id_spesialisasi SMALLINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id_spesialisasi INT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     nama_spesialisasi VARCHAR(20)
 );
 
 CREATE TABLE Ruang(
-    id_ruang SMALLINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
-    no_ruang SMALLINT UNIQUE NOT NULL
+    id_ruang INT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    no_ruang INT UNIQUE NOT NULL
 );
 
 -- P = Perempuan, L = Laki
@@ -39,7 +39,7 @@ CREATE TYPE GOL_DARAH AS ENUM ('A','B','AB','O');
 
 -- PASIEN
 CREATE TABLE Pasien(
-    id_pasien SMALLINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id_pasien INT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     nama VARCHAR(20) NOT NULL,
     no_telp CHAR(12) NOT NULL,
     email VARCHAR(20) NOT NULL UNIQUE,
@@ -49,11 +49,11 @@ CREATE TABLE Pasien(
     jenis_kelamin GENDER NOT NULL,
     tanggal_lahir DATE NOT NULL,
     is_active ACTIVE NOT NULL,
-    id_kelurahan SMALLINT REFERENCES Kelurahan(id_kelurahan) 
+    id_kelurahan INT REFERENCES Kelurahan(id_kelurahan) 
 );
 
 CREATE TABLE RekamMedis(
-    id_rkm_med SMALLINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id_rkm_med INT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     resep_obat VARCHAR(30)[] NOT NULL,
     prognosis_tindakan_lanjut VARCHAR(30)[] NOT NULL,
     diag_penunjang VARCHAR(30)[] NOT NULL,
@@ -61,20 +61,20 @@ CREATE TABLE RekamMedis(
     pemeriksaan_penunjang VARCHAR(30)[] NOT NULL,
     riwayat_penyakit VARCHAR(30)[] NOT NULL,
     keluhan VARCHAR(30)[] NOT NULL,
-    tinggi_badan SMALLINT NOT NULL,
-    berat_badan SMALLINT NOT NULL,
+    tinggi_badan INT NOT NULL,
+    berat_badan INT NOT NULL,
     golongan_darah GOL_DARAH NOT NULL,
-    diastolik SMALLINT NOT NULL,
-    sistolik SMALLINT NOT NULL,
-    id_pasien SMALLINT REFERENCES Pasien(id_pasien)
+    diastolik INT NOT NULL,
+    sistolik INT NOT NULL,
+    id_pasien INT REFERENCES Pasien(id_pasien)
 );
 
 
 CREATE TABLE DokumenRekamMedis (
-    id_dkm SMALLINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id_dkm INT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     uploaded_at DATE NOT NULL,
     path_file VARCHAR(30),
-    id_rkm_med SMALLINT REFERENCES RekamMedis(id_rkm_med)
+    id_rkm_med INT REFERENCES RekamMedis(id_rkm_med)
 );
 
 -- PEGAWAI
@@ -83,7 +83,7 @@ CREATE TABLE DokumenRekamMedis (
 CREATE TYPE ROLE_P AS ENUM ('Dokter', 'SisAdmin','PetAdmin','Perawat');
 
 CREATE TABLE Pegawai(
-    id_pegawai SMALLINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id_pegawai INT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     nama VARCHAR(20) NOT NULL,
     no_telp CHAR(12) NOT NULL,
     email VARCHAR(20) NOT NULL UNIQUE,
@@ -91,10 +91,10 @@ CREATE TABLE Pegawai(
     NIP CHAR(18) NOT NULL UNIQUE,
     created_at DATE NOT NULL,
     is_active ACTIVE NOT NULL,
-    id_kelurahan SMALLINT REFERENCES Kelurahan(id_kelurahan),
+    id_kelurahan INT REFERENCES Kelurahan(id_kelurahan),
     role ROLE_P NOT NULL,
     biaya_kunjungan INT NULL,
-    id_spesialisasi SMALLINT REFERENCES Spesialisasi(id_spesialisasi) NULL
+    id_spesialisasi INT REFERENCES Spesialisasi(id_spesialisasi) NULL
 );
 
 
@@ -102,34 +102,34 @@ CREATE TABLE Pegawai(
 CREATE TYPE STATUS_DAFTAR AS ENUM ('Pendaftaran','Pemanggilan','Dokter', 'Pemeriksaan','Tuntas');
 
 CREATE TABLE Pendaftaran (
-    id_pendaftaran SMALLINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id_pendaftaran INT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     status STATUS_DAFTAR NOT NULL,
     antrian INT NOT NULL,
     tanggal DATE NOT NULL,
-    id_pasien SMALLINT REFERENCES Pasien(id_pasien) NOT NULL
+    id_pasien INT REFERENCES Pasien(id_pasien) NOT NULL
 ); 
 
 CREATE TYPE HARI AS ENUM ('Senin','Selasa','Rabu','Kamis','Jumat','Sabtu','Minggu');
 
 CREATE TABLE JadwalPraktikDokter (
-    id_jadwal SMALLINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id_jadwal INT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     hari HARI NOT NULL,
     jam TIME NOT NULL,
-    kuota SMALLINT NOT NULL,
-    id_pegawai SMALLINT REFERENCES Pegawai(id_pegawai)NOT NULL,
-    id_ruang SMALLINT REFERENCES Ruang(id_ruang)
+    kuota INT NOT NULL,
+    id_pegawai INT REFERENCES Pegawai(id_pegawai)NOT NULL,
+    id_ruang INT REFERENCES Ruang(id_ruang)
 );
 
 CREATE TYPE METODE AS ENUM ('OVO','GOPAY','BCA');
 
 CREATE TABLE Transaksi(
-    id_transaksi SMALLINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id_transaksi INT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     metode METODE NULL,
     biaya_total INT NULL,
-    id_pendaftaran SMALLINT REFERENCES Pendaftaran(id_pendaftaran)
+    id_pendaftaran INT REFERENCES Pendaftaran(id_pendaftaran)
 );
 
 CREATE TABLE PendaftaranToJadwal(
-    id_pendaftaran SMALLINT REFERENCES Pendaftaran(id_pendaftaran) NOT NULL,
-    id_jadwal SMALLINT REFERENCES JadwalPraktikDokter(id_jadwal) NOT NULL
+    id_pendaftaran INT REFERENCES Pendaftaran(id_pendaftaran) NOT NULL,
+    id_jadwal INT REFERENCES JadwalPraktikDokter(id_jadwal) NOT NULL
 );
